+++ example-app/app/books/effects/collection.spec.ts
@@ -36,12 +36,12 @@ describe('CollectionEffects', () => {
         CollectionEffects,
         {
           provide: Database,
-          useValue: jasmine.createSpyObj('database', [
-            'open',
-            'query',
-            'insert',
-            'executeWrite',
-          ]),
+          useValue: {
+            open: jest.fn(),
+            query: jest.fn(),
+            insert: jest.fn(),
+            executeWrite: jest.fn(),
+          },
         },
         { provide: Actions, useFactory: getActions },
       ],
@@ -60,68 +60,68 @@ describe('CollectionEffects', () => {
   });
 
   describe('loadCollection$', () => {
-    it('should return a collection.LoadSuccessAction, with the books, on success', () => {
-      const action = new collection.LoadAction();
-      const completion = new collection.LoadSuccessAction([book1, book2]);
+    it('should return a collection.LoadSuccess, with the books, on success', () => {
+      const action = new collection.Load();
+      const completion = new collection.LoadSuccess([book1, book2]);
 
       actions$.stream = hot('-a', { a: action });
       const response = cold('-a-b|', { a: book1, b: book2 });
       const expected = cold('-----c', { c: completion });
-      db.query.and.returnValue(response);
+      db.query = jest.fn(() => response);
 
       expect(effects.loadCollection$).toBeObservable(expected);
     });
 
-    it('should return a collection.LoadFailAction, if the query throws', () => {
-      const action = new collection.LoadAction();
+    it('should return a collection.LoadFail, if the query throws', () => {
+      const action = new collection.Load();
       const error = 'Error!';
-      const completion = new collection.LoadFailAction(error);
+      const completion = new collection.LoadFail(error);
 
       actions$.stream = hot('-a', { a: action });
       const response = cold('-#', {}, error);
       const expected = cold('--c', { c: completion });
-      db.query.and.returnValue(response);
+      db.query = jest.fn(() => response);
 
       expect(effects.loadCollection$).toBeObservable(expected);
     });
   });
 
   describe('addBookToCollection$', () => {
-    it('should return a collection.AddBookSuccessAction, with the book, on success', () => {
-      const action = new collection.AddBookAction(book1);
-      const completion = new collection.AddBookSuccessAction(book1);
+    it('should return a collection.AddBookSuccess, with the book, on success', () => {
+      const action = new collection.AddBook(book1);
+      const completion = new collection.AddBookSuccess(book1);
 
       actions$.stream = hot('-a', { a: action });
       const response = cold('-b', { b: true });
       const expected = cold('--c', { c: completion });
-      db.insert.and.returnValue(response);
+      db.insert = jest.fn(() => response);
 
       expect(effects.addBookToCollection$).toBeObservable(expected);
       expect(db.insert).toHaveBeenCalledWith('books', [book1]);
     });
 
-    it('should return a collection.AddBookFailAction, with the book, when the db insert throws', () => {
-      const action = new collection.AddBookAction(book1);
-      const completion = new collection.AddBookFailAction(book1);
+    it('should return a collection.AddBookFail, with the book, when the db insert throws', () => {
+      const action = new collection.AddBook(book1);
+      const completion = new collection.AddBookFail(book1);
       const error = 'Error!';
 
       actions$.stream = hot('-a', { a: action });
       const response = cold('-#', {}, error);
       const expected = cold('--c', { c: completion });
-      db.insert.and.returnValue(response);
+      db.insert = jest.fn(() => response);
 
       expect(effects.addBookToCollection$).toBeObservable(expected);
     });
 
     describe('removeBookFromCollection$', () => {
-      it('should return a collection.RemoveBookSuccessAction, with the book, on success', () => {
-        const action = new collection.RemoveBookAction(book1);
-        const completion = new collection.RemoveBookSuccessAction(book1);
+      it('should return a collection.RemoveBookSuccess, with the book, on success', () => {
+        const action = new collection.RemoveBook(book1);
+        const completion = new collection.RemoveBookSuccess(book1);
 
         actions$.stream = hot('-a', { a: action });
         const response = cold('-b', { b: true });
         const expected = cold('--c', { c: completion });
-        db.executeWrite.and.returnValue(response);
+        db.executeWrite = jest.fn(() => response);
 
         expect(effects.removeBookFromCollection$).toBeObservable(expected);
         expect(db.executeWrite).toHaveBeenCalledWith('books', 'delete', [
@@ -129,15 +129,15 @@ describe('CollectionEffects', () => {
         ]);
       });
 
-      it('should return a collection.RemoveBookFailAction, with the book, when the db insert throws', () => {
-        const action = new collection.RemoveBookAction(book1);
-        const completion = new collection.RemoveBookFailAction(book1);
+      it('should return a collection.RemoveBookFail, with the book, when the db insert throws', () => {
+        const action = new collection.RemoveBook(book1);
+        const completion = new collection.RemoveBookFail(book1);
         const error = 'Error!';
 
         actions$.stream = hot('-a', { a: action });
         const response = cold('-#', {}, error);
         const expected = cold('--c', { c: completion });
-        db.executeWrite.and.returnValue(response);
+        db.executeWrite = jest.fn(() => response);
 
         expect(effects.removeBookFromCollection$).toBeObservable(expected);
         expect(db.executeWrite).toHaveBeenCalledWith('books', 'delete', [