+++ example-app/app/core/actions/layout.ts
@@ -3,12 +3,12 @@ import { Action } from '@ngrx/store';
 export const OPEN_SIDENAV = '[Layout] Open Sidenav';
 export const CLOSE_SIDENAV = '[Layout] Close Sidenav';
 
-export class OpenSidenavAction implements Action {
+export class OpenSidenav implements Action {
   readonly type = OPEN_SIDENAV;
 }
 
-export class CloseSidenavAction implements Action {
+export class CloseSidenav implements Action {
   readonly type = CLOSE_SIDENAV;
 }
 
-export type Actions = OpenSidenavAction | CloseSidenavAction;
+export type Actions = OpenSidenav | CloseSidenav;