+++ example-app/app/reducers/index.ts
@@ -3,8 +3,18 @@ import {
   createSelector,
   createFeatureSelector,
   ActionReducer,
+  MetaReducer,
 } from '@ngrx/store';
 import { environment } from '../../environments/environment';
+import { RouterStateUrl } from '../shared/utils';
+import * as fromRouter from '@ngrx/router-store';
+
+/**
+ * storeFreeze prevents state from being mutated. When mutation occurs, an
+ * exception will be thrown. This is useful during development mode to
+ * ensure that none of the reducers accidentally mutates the state.
+ */
+import { storeFreeze } from 'ngrx-store-freeze';
 
 /**
  * Every reducer module's default export is the reducer function itself. In
@@ -21,6 +31,7 @@ import * as fromLayout from '../core/reducers/layout';
  */
 export interface State {
   layout: fromLayout.State;
+  routerReducer: fromRouter.RouterReducerState<RouterStateUrl>;
 }
 
 /**
@@ -30,10 +41,11 @@ export interface State {
  */
 export const reducers: ActionReducerMap<State> = {
   layout: fromLayout.reducer,
+  routerReducer: fromRouter.routerReducer,
 };
 
 // console.log all actions
-export function logger(reducer: ActionReducer<State>): ActionReducer<any, any> {
+export function logger(reducer: ActionReducer<State>): ActionReducer<State> {
   return function(state: State, action: any): State {
     console.log('state', state);
     console.log('action', action);
@@ -47,8 +59,8 @@ export function logger(reducer: ActionReducer<State>): ActionReducer<any, any> {
  * the root meta-reducer. To add more meta-reducers, provide an array of meta-reducers
  * that will be composed to form the root meta-reducer.
  */
-export const metaReducers: ActionReducer<any, any>[] = !environment.production
-  ? [logger]
+export const metaReducers: MetaReducer<State>[] = !environment.production
+  ? [logger, storeFreeze]
   : [];
 
 /**