+++ modules/effects/src/effects_feature_module.ts
@@ -1,4 +1,5 @@
-import { NgModule, Inject, Type } from '@angular/core';
+import { NgModule, Inject, Optional } from '@angular/core';
+import { StoreModule } from '@ngrx/store';
 import { EffectsRootModule } from './effects_root_module';
 import { FEATURE_EFFECTS } from './tokens';
 
@@ -6,7 +7,8 @@ import { FEATURE_EFFECTS } from './tokens';
 export class EffectsFeatureModule {
   constructor(
     private root: EffectsRootModule,
-    @Inject(FEATURE_EFFECTS) effectSourceGroups: any[][]
+    @Inject(FEATURE_EFFECTS) effectSourceGroups: any[][],
+    @Optional() storeModule: StoreModule
   ) {
     effectSourceGroups.forEach(group =>
       group.forEach(effectSourceInstance =>