+++ modules/effects/src/effects_root_module.ts
@@ -1,4 +1,5 @@
-import { NgModule, Inject } from '@angular/core';
+import { NgModule, Inject, Optional } from '@angular/core';
+import { StoreModule } from '@ngrx/store';
 import { EffectsRunner } from './effects_runner';
 import { EffectSources } from './effect_sources';
 import { ROOT_EFFECTS } from './tokens';
@@ -8,7 +9,8 @@ export class EffectsRootModule {
   constructor(
     private sources: EffectSources,
     runner: EffectsRunner,
-    @Inject(ROOT_EFFECTS) rootEffects: any[]
+    @Inject(ROOT_EFFECTS) rootEffects: any[],
+    @Optional() storeModule: StoreModule
   ) {
     runner.start();