+++ modules/effects/src/index.ts
@@ -5,3 +5,4 @@ export { EffectsModule } from './effects_module';
 export { EffectSources } from './effect_sources';
 export { OnRunEffects } from './on_run_effects';
 export { toPayload } from './util';
+export { EffectNotification } from './effect_notification';