+++ modules/router-store/spec/integration.spec.ts
@@ -1,6 +1,6 @@
-import { Component } from '@angular/core';
+import { Component, Provider } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { NavigationEnd, Router } from '@angular/router';
+import { NavigationEnd, Router, RouterStateSnapshot } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Store, StoreModule } from '@ngrx/store';
 import {
@@ -10,13 +10,17 @@ import {
   RouterAction,
   routerReducer,
   StoreRouterConnectingModule,
+  RouterStateSerializer,
 } from '../src/index';
 import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/first';
+import 'rxjs/add/operator/mapTo';
+import 'rxjs/add/operator/take';
 import 'rxjs/add/operator/toPromise';
+import { of } from 'rxjs/observable/of';
 
 describe('integration spec', () => {
-  it('should work', done => {
+  it('should work', (done: any) => {
     const reducer = (state: string = '', action: RouterAction<any>) => {
       if (action.type === ROUTER_NAVIGATION) {
         return action.payload.routerState.url.toString();
@@ -58,7 +62,7 @@ describe('integration spec', () => {
       });
   });
 
-  it('should support preventing navigation', done => {
+  it('should support preventing navigation', (done: any) => {
     const reducer = (state: string = '', action: RouterAction<any>) => {
       if (
         action.type === ROUTER_NAVIGATION &&
@@ -94,7 +98,7 @@ describe('integration spec', () => {
       });
   });
 
-  it('should support rolling back if navigation gets canceled', done => {
+  it('should support rolling back if navigation gets canceled', (done: any) => {
     const reducer = (state: string = '', action: RouterAction<any>): any => {
       if (action.type === ROUTER_NAVIGATION) {
         return {
@@ -152,7 +156,7 @@ describe('integration spec', () => {
       });
   });
 
-  it('should support rolling back if navigation errors', done => {
+  it('should support rolling back if navigation errors', (done: any) => {
     const reducer = (state: string = '', action: RouterAction<any>): any => {
       if (action.type === ROUTER_NAVIGATION) {
         return {
@@ -212,7 +216,9 @@ describe('integration spec', () => {
       });
   });
 
-  it('should call navigateByUrl when resetting state of the routerReducer', done => {
+  it('should call navigateByUrl when resetting state of the routerReducer', (
+    done: any
+  ) => {
     const reducer = (state: any, action: RouterAction<any>) => {
       const r = routerReducer(state, action);
       return r && r.state
@@ -286,10 +292,134 @@ describe('integration spec', () => {
         done();
       });
   });
+
+  it('should support cancellation of initial navigation using canLoad guard', (
+    done: any
+  ) => {
+    const reducer = (state: any, action: RouterAction<any>) => {
+      const r = routerReducer(state, action);
+      return r && r.state
+        ? { url: r.state.url, navigationId: r.navigationId }
+        : null;
+    };
+
+    createTestModule({
+      reducers: { routerReducer, reducer },
+      canLoad: () => false,
+    });
+
+    const router = TestBed.get(Router);
+    const store = TestBed.get(Store);
+    const log = logOfRouterAndStore(router, store);
+
+    router.navigateByUrl('/load').then((r: boolean) => {
+      expect(r).toBe(false);
+
+      expect(log).toEqual([
+        { type: 'store', state: null },
+        { type: 'router', event: 'NavigationStart', url: '/load' },
+        { type: 'store', state: null },
+        { type: 'router', event: 'NavigationCancel', url: '/load' },
+      ]);
+      done();
+    });
+
+    it('should support a custom RouterStateSnapshot serializer ', (
+      done: any
+    ) => {
+      const reducer = (state: any, action: RouterAction<any>) => {
+        const r = routerReducer(state, action);
+        return r && r.state
+          ? { url: r.state.url, navigationId: r.navigationId }
+          : null;
+      };
+
+      class CustomSerializer
+        implements RouterStateSerializer<{ url: string; params: any }> {
+        serialize(routerState: RouterStateSnapshot) {
+          const url = `${routerState.url}-custom`;
+          const params = { test: 1 };
+
+          return { url, params };
+        }
+      }
+
+      const providers = [
+        { provide: RouterStateSerializer, useClass: CustomSerializer },
+      ];
+
+      createTestModule({ reducers: { routerReducer, reducer }, providers });
+
+      const router = TestBed.get(Router);
+      const store = TestBed.get(Store);
+      const log = logOfRouterAndStore(router, store);
+
+      router
+        .navigateByUrl('/')
+        .then(() => {
+          log.splice(0);
+          return router.navigateByUrl('next');
+        })
+        .then(() => {
+          expect(log).toEqual([
+            { type: 'router', event: 'NavigationStart', url: '/next' },
+            { type: 'router', event: 'RoutesRecognized', url: '/next' },
+            {
+              type: 'store',
+              state: {
+                url: '/next-custom',
+                navigationId: 2,
+                params: { test: 1 },
+              },
+            },
+            { type: 'router', event: 'NavigationEnd', url: '/next' },
+          ]);
+          log.splice(0);
+          done();
+        });
+    });
+  });
+
+  it('should support event during an async canActivate guard', (done: any) => {
+    createTestModule({
+      reducers: { routerReducer },
+      canActivate: () => {
+        store.dispatch({ type: 'USER_EVENT' });
+        return store.take(1).mapTo(true);
+      },
+    });
+
+    const router: Router = TestBed.get(Router);
+    const store: Store<any> = TestBed.get(Store);
+    const log = logOfRouterAndStore(router, store);
+
+    router
+      .navigateByUrl('/')
+      .then(() => {
+        log.splice(0);
+        return router.navigateByUrl('next');
+      })
+      .then(() => {
+        expect(log).toEqual([
+          { type: 'router', event: 'NavigationStart', url: '/next' },
+          { type: 'router', event: 'RoutesRecognized', url: '/next' },
+          { type: 'store', state: undefined }, // after ROUTER_NAVIGATION
+          { type: 'store', state: undefined }, // after USER_EVENT
+          { type: 'router', event: 'NavigationEnd', url: '/next' },
+        ]);
+
+        done();
+      });
+  });
 });
 
 function createTestModule(
-  opts: { reducers?: any; canActivate?: Function } = {}
+  opts: {
+    reducers?: any;
+    canActivate?: Function;
+    canLoad?: Function;
+    providers?: Provider[];
+  } = {}
 ) {
   @Component({
     selector: 'test-app',
@@ -314,6 +444,11 @@ function createTestModule(
           component: SimpleCmp,
           canActivate: ['CanActivateNext'],
         },
+        {
+          path: 'load',
+          loadChildren: 'test',
+          canLoad: ['CanLoadNext'],
+        },
       ]),
       StoreRouterConnectingModule,
     ],
@@ -322,6 +457,11 @@ function createTestModule(
         provide: 'CanActivateNext',
         useValue: opts.canActivate || (() => true),
       },
+      {
+        provide: 'CanLoadNext',
+        useValue: opts.canLoad || (() => true),
+      },
+      opts.providers || [],
     ],
   });