+++ modules/store/src/reducer_manager.ts
@@ -40,7 +40,7 @@ export class ReducerManager extends BehaviorSubject<ActionReducer<any, any>>
   }: StoreFeature<any, any>) {
     const reducer =
       typeof reducers === 'function'
-        ? reducers
+        ? (state: any, action: any) => reducers(state || initialState, action)
         : createReducerFactory(reducerFactory, metaReducers)(
             reducers,
             initialState