+++ modules/store/src/utils.ts
@@ -3,6 +3,7 @@ import {
   ActionReducer,
   ActionReducerMap,
   ActionReducerFactory,
+  MetaReducer,
 } from './models';
 
 export function combineReducers<T, V extends Action = Action>(
@@ -25,7 +26,8 @@ export function combineReducers(
 
   const finalReducerKeys = Object.keys(finalReducers);
 
-  return function combination(state = initialState, action) {
+  return function combination(state, action) {
+    state = state || initialState;
     let hasChanged = false;
     const nextState: any = {};
     for (let i = 0; i < finalReducerKeys.length; i++) {
@@ -84,10 +86,10 @@ export function compose(...functions: any[]) {
   };
 }
 
-export function createReducerFactory(
-  reducerFactory: ActionReducerFactory<any, any>,
-  metaReducers?: ActionReducer<any, any>[]
-): ActionReducerFactory<any, any> {
+export function createReducerFactory<T, V extends Action = Action>(
+  reducerFactory: ActionReducerFactory<T, V>,
+  metaReducers?: MetaReducer<T, V>[]
+): ActionReducerFactory<T, V> {
   if (Array.isArray(metaReducers) && metaReducers.length > 0) {
     return compose.apply(null, [...metaReducers, reducerFactory]);
   }